+++ integrations/grpcloglog/logger_test.go
@@ -1,0 +1,12 @@
+package grpcloglog
+
+import (
+	"testing"
+
+	"github.com/goph/logur"
+	"google.golang.org/grpc/grpclog"
+)
+
+func TestLogger(t *testing.T) {
+	var _ grpclog.LoggerV2 = New(logur.NewNoopLogger())
+}